backend/app/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
orjson==3.10.18
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
ujson==5.10.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
